requirements.txt
yfinance>=0.2.36
vectorbt>=0.26.1
pandas-ta>=0.3.14b
streamlit>=1.32.0
plotly>=5.20.0
duckdb>=0.10.0
fastapi>=0.110.0
uvicorn>=0.29.0
openai>=1.14.0
python-dotenv>=1.0.1
pandas>=2.0.0
numpy>=1.26.0
pyarrow>=15.0.0
alpaca-py>=0.20.0
requests>=2.31.0
